Break workspace tests into their own file
